TINY-3901: move RemoveTrailingBr.ts loacation to `modules/tinymce/src/core/main/ts/dom`
