+++ modules/tinymce/src/core/main/ts/dom/DomSerializerFilters.ts
@@ -4,7 +4,7 @@ import DOMUtils from '../api/dom/DOMUtils';
 import DomParser from '../api/html/DomParser';
 import Entities from '../api/html/Entities';
 import AstNode from '../api/html/Node';
-import * as RemoveTrailingBr from '../RemoveTrailingBr';
+import * as RemoveTrailingBr from './RemoveTrailingBr';
 import * as Zwsp from '../text/Zwsp';
 import { DomSerializerSettings } from './DomSerializerImpl';
