+++ modules/tinymce/src/core/main/ts/dom/RemoveTrailingBr.ts
@@ -1,11 +1,11 @@
 import { Unicode } from '@ephox/katamari';
 
-import DomParser from './api/html/DomParser';
-import AstNode from './api/html/Node';
-import Schema from './api/html/Schema';
-import Tools from './api/util/Tools';
-import * as TransparentElements from './content/TransparentElements';
-import { isEmpty, paddEmptyNode } from './html/ParserUtils';
+import DomParser from '../api/html/DomParser';
+import AstNode from '../api/html/Node';
+import Schema from '../api/html/Schema';
+import Tools from '../api/util/Tools';
+import * as TransparentElements from '../content/TransparentElements';
+import { isEmpty, paddEmptyNode } from '../html/ParserUtils';
 
 export const addNodeFilter = (htmlParser: DomParser, schema: Schema): void => {
   htmlParser.addNodeFilter('br', (nodes, _, args) => {
